完善账户信息窗口的功能 新增复制账户名称和密码到剪贴板的功能 新增复制到剪贴板时的音效 修复工具栏打开URL按钮无效的bug

+++ main/Functional.cpp
@@ -249,7 +249,7 @@ void AccountType::assignTo( winplus::Mixed * accountTypeMixed, CString const & f
 
 void AccountType::assign( winplus::Mixed const & accountTypeMixed )
 {
-    int n = accountTypeMixed.getCount();
+    int n = (int)accountTypeMixed.getCount();
     for ( int i = 0; i < n; ++i )
     {
         auto & pr = accountTypeMixed.getPair(i);
@@ -419,7 +419,7 @@ void AccountCate::assignTo( winplus::Mixed * accountCateMixed, CString const & f
 
 void AccountCate::assign( winplus::Mixed const & accountCateMixed )
 {
-    int n = accountCateMixed.getCount();
+    int n = (int)accountCateMixed.getCount();
     for ( int i = 0; i < n; ++i )
     {
         auto & pr = accountCateMixed.getPair(i);
@@ -681,7 +681,7 @@ void Account::assignTo( winplus::Mixed * accountMixed, CString const & fieldName
 
 void Account::assign( winplus::Mixed const & accountMixed )
 {
-    int n = accountMixed.getCount();
+    int n = (int)accountMixed.getCount();
     for ( int i = 0; i < n; ++i )
     {
         auto & pr = accountMixed.getPair(i);
